Custom var defines

+++ Lila/src/Lila/Utils/UUID.cpp
@@ -6,7 +6,7 @@ namespace Lila {
 
 	static std::random_device device;
 	static std::mt19937_64 engine(device());
-	static std::uniform_int_distribution<unsigned long long> uniformDist;
+	static std::uniform_int_distribution<u64> uniformDist;
 
 	UUID::UUID() : id(uniformDist(engine)) {}
